+++ 30_sources/SRC_Zupfnoter/src/opal-ajv.rb
@@ -40,20 +40,21 @@ module Ajv
       {:$schema     => "http://json-schema.org/draft-04/schema#",
        :description => "Generated from x.json with shasum 0b1781e0803dc084178858e9fbe2b4e0b65c08e7",
        :type        => "object",
-       :required    => ["confstack", "produce", "abc_parser", "restposition", "wrap", "defaults", "templates", "annotations", "extract", "layout", "neatjson"],
+       :required    => ["produce", "abc_parser", "restposition", "wrap", "defaults", "templates", "annotations", "extract", "layout", "neatjson"],
 
        :definitions => {
-           :pos           => {:type        => "array",
-                              :minItems    => 2,
-                              :uniqueItems => false,
-                              :items       => {:type => "number"}},
-           :notes_entry   => {:type       => "object",
-                              :required   => ["pos", "text", "style"],
-                              :properties =>
-                                  {:pos   => {:'$ref' => '#/definitions/pos'},
-                                   :text  => {:type => "string"},
-                                   :style => {:type => "string"}}},
-           :minc_entry => {
+
+           :pos         => {:type        => "array",
+                            :minItems    => 2,
+                            :uniqueItems => false,
+                            :items       => {:type => "number"}},
+           :notes_entry => {:type       => "object",
+                            :required   => ["pos", "text", "style"],
+                            :properties =>
+                                {:pos   => {:'$ref' => '#/definitions/pos'},
+                                 :text  => {:type => "string"},
+                                 :style => {:type => "string"}}},
+           :minc_entry  => {
                :type                 => "object",
                :required             => [:minc_f],
                :additionalProperties => false,
@@ -232,6 +233,12 @@ module Ajv
                                                                     :additionalProperties => false,
                                                                     :properties           =>
                                                                         {:limit_a3          => {:type => "boolean"},
+                                                                         :jumpline_anchor   => {:'$ref' => '#/definitions/pos'},
+                                                                         :manual_sheet      => {:type       => 'object',
+                                                                                                required:   [:llpos, :trpos, :url],
+                                                                                                :properties => {llpos: {:'$ref' => '#/definitions/pos'},
+                                                                                                                trpos: {:'$ref' => '#/definitions/pos'},
+                                                                                                                url:   {:type => 'string'}}},
                                                                          :LINE_THIN         => {:type => "number"},
                                                                          :LINE_MEDIUM       => {:type => "number"},
                                                                          :LINE_THICK        => {:type => "number"},
@@ -245,6 +252,12 @@ module Ajv
                                                                                                 :minItems => 2,
                                                                                                 :items    => {:type => "number"}},
                                                                          :grid              => {:type => "boolean"},
+                                                                         :jumpline_anchor   => {:"$ref" => "#/definitions/pos"},
+                                                                         :color             => {:type       => 'object',
+                                                                                                :properties => {:color_default  => {:type => 'string'},
+                                                                                                                :color_variant1 => {:type => 'string'},
+                                                                                                                :color_variant2 => {:type => 'string'}
+                                                                                                }},
                                                                          :packer            => {:type       => 'object',
                                                                                                 :properties => {
                                                                                                     :pack_method            => {:type => 'integer'},